+++ NEWS.md
@@ -1,6 +1,7 @@
 # formods 0.1.2
 
 * development version
+* Added `new_module_template()` to create new module templates. 
 
 # formods 0.1.1